Extract table selection highlighting.
